requirements.txt
open3d==0.19.0
numpy==1.24.4
trimesh==4.1.3
torch==2.2.0
torch-geometric==2.4.0
pandas==2.2.0
biopython==1.83
